Extract WhenWhereCard so the gedu body shows day/time too
The gedu detail body shipped without the day/time/location/age/language
panel the parent body uses. Lifting it into its own file lets both bodies
reuse a single source of truth — same row shape (ProductV2BrowseRow),
same layout — so the gedu page no longer hides info the parent page
already renders the same way.
