+++ src/components/public/products-v2/product-detail-page-body.tsx
@@ -2,10 +2,9 @@
 
 import Link from "next/link";
 import { useLocale, useTranslations } from "next-intl";
-import { ArrowLeft, Clock, Globe, MapPin, Sparkles, Users } from "lucide-react";
+import { ArrowLeft } from "lucide-react";
 import { Card, CardContent } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
-import { LanguageFlag } from "@/components/ui/language-flag";
 import { ProductThumbnail } from "@/components/ui/product-thumbnail";
 import { resolveLocale } from "@/lib/constants/locales";
 import { resolveTranslation } from "@/lib/i18n/resolve-translation";
@@ -17,11 +16,7 @@ import type {
 import { formatInTimeZone } from "date-fns-tz";
 import { computeProductSessions } from "@/components/calendar/compute-product-sessions";
 import { SessionCalendarView } from "@/components/calendar/session-calendar-view";
-import { formatProductLocation } from "./format-product-location";
-import {
-  formatProductSchedule,
-  type ProductScheduleSummary,
-} from "./format-product-schedule";
+import { ProductWhenWhereCard } from "./product-when-where-card";
 import type { RegistrationState } from "./derive-registration-state";
 import { SignupPanel } from "./signup-panel";
 import type { AuthState, MyParticipationState } from "./signup-panel-view";
@@ -175,51 +170,10 @@ function MainColumn({
   const uiLocale = resolveLocale(useLocale());
 
   const tagLabels = resolveTagLabels(product.product_tags_v2, uiLocale);
-  const schedule = formatProductSchedule({ product, locale: uiLocale });
-  const scheduleLines = renderScheduleLinesForDetail(schedule);
-  const location = formatProductLocation(product);
 
   return (
     <div className="space-y-6">
-      <Card>
-        <CardContent className="space-y-3 p-5 sm:p-6 text-sm">
-          <h2 className="text-sm font-semibold uppercase tracking-wider text-muted-foreground">
-            {t("sections.whenWhere")}
-          </h2>
-          <DetailRow icon={Clock} label={t("info.schedule")}>
-            {scheduleLines.length === 1 ? (
-              scheduleLines[0]
-            ) : (
-              <ul className="space-y-0.5">
-                {scheduleLines.map((line, idx) => (
-                  <li key={idx}>{line}</li>
-                ))}
-              </ul>
-            )}
-          </DetailRow>
-          <DetailRow
-            icon={product.is_remote && location?.kind !== "muni" ? Globe : MapPin}
-            label={
-              product.is_remote && location?.kind !== "muni"
-                ? t("info.format")
-                : t("info.where")
-            }
-          >
-            {renderLocationLine({
-              location,
-              isRemote: product.is_remote,
-              tOnline: t("info.online"),
-              tTbd: t("info.tbd"),
-            })}
-          </DetailRow>
-          <DetailRow icon={Users} label={t("info.ageRange")}>
-            {t("info.ages", { min: product.min_age, max: product.max_age })}
-          </DetailRow>
-          <DetailRow icon={Sparkles} label={t("info.language")}>
-            <LanguageFlag code={product.spoken_language_code} />
-          </DetailRow>
-        </CardContent>
-      </Card>
+      <ProductWhenWhereCard product={product} />
 
       <CalendarCard product={product} />
 
@@ -303,89 +257,6 @@ function CalendarCard({
       </CardContent>
     </Card>
   );
-}
-
-function DetailRow({
-  icon: Icon,
-  label,
-  children,
-}: {
-  icon: React.ComponentType<{ className?: string }>;
-  label: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="flex gap-3">
-      <Icon className="mt-0.5 h-4 w-4 shrink-0 text-muted-foreground" />
-      <div className="min-w-0">
-        <dt className="text-xs text-muted-foreground">{label}</dt>
-        <dd className="mt-0.5">{children}</dd>
-      </div>
-    </div>
-  );
-}
-
-function renderLocationLine({
-  location,
-  isRemote,
-  tOnline,
-  tTbd,
-}: {
-  location: ReturnType<typeof formatProductLocation>;
-  isRemote: boolean;
-  tOnline: string;
-  tTbd: string;
-}): string {
-  if (!location) return isRemote ? tOnline : tTbd;
-  switch (location.kind) {
-    case "site":
-      return location.parent
-        ? `${location.site}, ${location.parent}`
-        : location.site;
-    case "muni":
-      return location.name;
-  }
-}
-
-// Detail page has the room to break per-time-group lines apart, so a
-// multi-day camp ("Apr 30 – May 5" + "Mon, Wed, Fri · 09:00–15:00") and a
-// rare multi-time club (one line per group) both surface their full
-// shape here, with the card carrying the typical-case summary.
-//
-// Timezone label appears once on the lead line — repeating it on each
-// group reads as noise; readers infer same-tz for the whole schedule.
-function renderScheduleLinesForDetail(
-  schedule: ProductScheduleSummary,
-): string[] {
-  switch (schedule.kind) {
-    case "tbd":
-      return ["—"];
-    case "recurring":
-      return schedule.groups.map((g, idx) => {
-        const line = `${g.weekdaysLabel} · ${g.startTime}–${g.endTime}`;
-        return idx === 0 ? withTz(line, schedule.tz) : line;
-      });
-    case "ranged": {
-      const dateLine = withTz(
-        `${schedule.startDate} – ${schedule.endDate}`,
-        schedule.tz,
-      );
-      const groupLines = schedule.groups.map(
-        (g) => `${g.weekdaysLabel} · ${g.startTime}–${g.endTime}`,
-      );
-      return [dateLine, ...groupLines];
-    }
-    case "single": {
-      const time = schedule.time
-        ? ` · ${schedule.time.start}–${schedule.time.end}`
-        : "";
-      return [withTz(`${schedule.date}${time}`, schedule.tz)];
-    }
-  }
-}
-
-function withTz(line: string, tz: string): string {
-  return tz ? `${line} (${tz})` : line;
 }
 
 function resolveTagLabels(
